requirements.txt
asgiref==3.8.1
# backports.zoneinfo==0.2.1
bitstring==4.0.2
CacheControl==0.13.1
certifi==2023.5.7
cffi==1.15.1
channels==4.1.0
charset-normalizer==3.1.0
click==8.0.4
colorama==0.4.6
construct==2.10.68
contextlib2==21.6.0
cryptography==39.0.2
Django==4.2.14
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
ecdsa==0.18.0
esp-coredump==1.5.1
esp-idf-kconfig==1.1.0
esp-idf-monitor==1.1.1
esp-idf-size==0.3.1
esptool==4.6.2
filelock==3.12.2
freertos-gdb==1.0.2
future==0.18.2
gcovr==6.0
gunicorn==22.0.0
idf-component-manager==1.2.3
idna==3.4
Jinja2==3.0.3
kconfiglib==14.1.0
lxml==4.9.2
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
# pkg_resources==0.0.0
psutil==5.9.5
pycparser==2.21
pyelftools==0.29
pygdbmi==0.9.0.2
Pygments==2.15.1
pyparsing==3.0.9
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0
reedsolo==1.7.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
schema==0.7.5
six==1.16.0
sqlparse==0.5.1
termux==1.0.0
tqdm==4.65.0
typing_extensions==4.12.2
urllib3==1.26.16
websocket-client==1.6.1
xmlrunner==1.7.7
